component node化
