+++ sciurus17_control/package.xml
@@ -18,6 +18,7 @@
   <depend>hardware_interface</depend>
   <depend>pluginlib</depend>
   <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
   <depend>ros2_controllers</depend>
   <depend>rt_manipulators_cpp</depend>
