+++ sciurus17_examples/launch/composition.launch.py
@@ -1,0 +1,30 @@
+import launch
+from launch_ros.actions import ComposableNodeContainer
+from launch_ros.descriptions import ComposableNode
+
+
+def generate_launch_description():
+    """Generate launch description with multiple components."""
+    container = ComposableNodeContainer(
+            name='my_container',
+            namespace='',
+            package='rclcpp_components',
+            executable='component_container',
+            composable_node_descriptions=[
+                ComposableNode(
+                    package='sciurus17_examples',
+                    plugin='sciurus17_examples::ColorDetection',
+                    name='color_detection'),
+                ComposableNode(
+                    package='sciurus17_examples',
+                    plugin='sciurus17_examples::ObjectTracker',
+                    name='object_tracker'),
+                ComposableNode(
+                    package='sciurus17_examples',
+                    plugin='sciurus17_examples::NeckJtControl',
+                    name='neck_jt_control'),
+            ],
+            output='screen',
+    )
+
+    return launch.LaunchDescription([container])
